Rename pkg "frequent_issue_detection" -> "frequentissuedetection"
